Anthropic/requirements.txt
anthropic==0.19.1
python-dotenv==1.0.1